pyproject.toml
[tool.poetry]
name = "broker"
version = "0.1.0"
description = ""
authors = ["Sipan Margaryan <sipanm19@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
python-dotenv = "^1.0.0"
cachecontrol = "^0.12.7"
requests = "^2.29.0"
python-multipart = "^0.0.6"
django = "^5.0.1"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.1"
django-filter = "^23.5"
drf-yasg = "^1.21.7"
pytest-django = "^4.7.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
pytest-mock = "3.10.0"
factory-boy = "^3.3.0"
httpx = "^0.24.0"


[tool.isort]
extend_skip = []

[tool.black]
extend_exclude = ""

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
